requirements.txt
asgiref==3.4.1
Django==3.2.9
djangorestframework==3.12.4
greenlet==1.1.2
psycopg2==2.8.4
pytz==2021.3
SQLAlchemy==1.4.26
sqlparse==0.4.2
